COMMERCE-1650 Fixed commerce product definition local service implementation

+++ commerce-product-service/src/main/java/com/liferay/commerce/product/service/impl/CPDefinitionLocalServiceImpl.java
@@ -574,7 +574,12 @@ public class CPDefinitionLocalServiceImpl
 
 		// Commerce product
 
-		cProductLocalService.deleteCProduct(cpDefinition.getCProductId());
+		int cpDefinitionsCount = cpDefinitionPersistence.countByC_S(
+			cpDefinition.getCProductId(), WorkflowConstants.STATUS_ANY);
+
+		if (_isVersioningEnabled() && (cpDefinitionsCount == 1)) {
+			cProductLocalService.deleteCProduct(cpDefinition.getCProductId());
+		}
 
 		// Commerce product definition localization
